requirements.txt
faiss-cpu
google-genai
numpy
python-dotenv
requests
pydantic